+++ packages/app/src/pages/debug/atlas.tsx
@@ -11,6 +11,8 @@ import {
 
 import { GlyphControls } from '../../ui/glyph-controls';
 
+console.log([..."❤️‍🔥"].map(s => s.codePointAt(0).toString(16)))
+
 export const DebugAtlasPage: LC = () => {
 
   const view = (
@@ -42,9 +44,10 @@ export const DebugAtlasPage: LC = () => {
                     size={60}
                     snap={false}
                     family="Lato, Noto Emoji"
-                    text={" is ⭐️✨✌️🌿🌳🐲💧🍅🍲🫕🏀🏈 presented which allows improved rendering of glyphs composed of curved and linear elements. A distance field is generated from a high resolution image, and then stored into a channel of a lower-resolution texture.\n\nIn the simplest case, this texture can then be rendered simply by using the alpha-testing and alpha-thresholding feature of modern GPUs, without a custom shader. This allows the technique to be used on even the lowest-end 3D graphics hardware."}
+                    text={" is ⭐️✨✌️🌿🌳🐲🐬🍅🍲🫕🏀🏈🏴‍☠️👨‍🚀❤️‍🔥 presented which allows improved rendering of glyphs composed of curved and linear elements. A distance field is generated from a high resolution image, and then stored into a channel of a lower-resolution texture.\n\nIn the simplest case, this texture can then be rendered simply by using the alpha-testing and alpha-thresholding feature of modern GPUs, without a custom shader. This allows the technique to be used on even the lowest-end 3D graphics hardware."}
                     color={[0.5, 0.5, 0.5, 1]}
                   />
+                 {/*    */}
                 </Inline>
               </Absolute>
             </Layout>