+++ packages/workbench/src/text/font-loader.ts
@@ -1,5 +1,5 @@
 import type { LiveComponent, LiveElement } from '@use-gpu/live';
-import type { Font } from '@use-gpu/glyph';
+import type { Font, LazyFontSource } from '@use-gpu/glyph';
 
 import { use, gather, keyed, yeet, useOne } from '@use-gpu/live';
 import { toHash } from '@use-gpu/state';
@@ -11,7 +11,9 @@ export type FontSource = {
   family: string,
   weight: string | number,
   style: string,
-  src: string,
+
+  src?: string,
+  lazy?: LazyFontSource,
 };
 
 export type FontLoaderProps = {
@@ -23,19 +25,32 @@ const NO_FONTS: FontSource[] = [];
 
 export const FontLoader: LiveComponent<FontLoaderProps> = ({fonts = NO_FONTS, children}) => {
 
-  const resources = useOne(() => fonts
-    .filter((s: FontSource) => !!s.src)
-    .map((source: FontSource) =>
-      keyed(Fetch, toHash(source), {
-        url: source.src,
-        type: 'arrayBuffer',
-        render: (buffer: ArrayBuffer) => yeet({props: {
-          ...source,
-          weight: parseWeight(source.weight),
-        }, buffer}),
-      })
-    ),
-    fonts);
+  const resources = useOne(
+    () => fonts
+      .map((source: FontSource) => {
+        const key = toHash(source);
+        
+        const {family, style, weight} = source;
+        const props = {
+          family,
+          style,
+          weight: parseWeight(weight),
+        };
+
+        const {src, lazy} = source;
+        if (src != null) {
+          return keyed(Fetch, key, {
+            url: src,
+            type: 'arrayBuffer',
+            render: (buffer: ArrayBuffer) => yeet({ props, buffer }),
+          })
+        }
+        if (lazy != null) {
+          return yeet({ props, lazy });
+        }
+      }),
+    fonts,
+  );
 
   return gather(resources, (fonts: Font[]) => {
     return use(FontProvider, { fonts, children })