+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.9.3
+- `@use-gpu/glyph`: Support loading emoji PNGs on-demand
+- `@use-gpu/layout`: Add `aspect` to elements for sizing w/h relative to each other
+- `@use-gpu/plot`: Add UVs and STs to `<Surface>`
+
 0.9.2
 - `@use-gpu/inspect`: Fix inspecting of large typed arrays
 - `@use-gpu/present`: Add `<Overlay>` for independent overlays